requirements.txt
pandas
numpy
pynput
brainflow
pyserial
